[Segment Cache] Move cache key creation to client
When I originally set up the Segment Cache transport types, I had the
notion that the client should never have to compute the request key that
is sent to the server. That way we could, say, encrypt the keys on the
server to make them unguessable.

But once we added the concept of access tokens for that purpose, this
motivation was gone. The remaining motivation was to keep the contract
as simple as possible.

However, there's a lot of existing code that relies on manipulating the
"decoded" form of the segment values. Since we're sending the full
segment value anyway, it's wasteful to also send the encoded form, given
that it can be trivially re-derived by the client.

My immediate motivation is that, I need to be able to construct a cache
key from a FlightRouterState. As a bonus, this will also get us closer
to being able to remove the dynamic param values from the server
response, a separate goal that we're currently working on.

We can revisit this once if/when we're able to remove the existing cases
where the full segment is needed by the client.